Add files via upload
The predictors of in-hospital mortality for admitted patients remain poorly characterized. The  aim  of this project is to develop and validate a prediction model for all-cause in-hospital mortality among admitted patients, and identify factors that are most relevant to the ICU survival rate.
The project aims to perform analysis that can provide clinical research and public health professionals with valuable insights, save lives by increasing patients’ survival rates, and ultimately bring a positive impact to the community.